+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "fc39c5a69f2544fdb01796463330d5af434a8073",
+  "pnpmShrinkwrapHash": "7a000515f12cd1697a48fafd964e66b9cc6812d8",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
