+++ packages/apps/kadena-docs/src/components/Layout/Full/Full.tsx
@@ -1,21 +1,14 @@
-import { styled } from '@kadena/react-components';
+import { Heading, styled } from '@kadena/react-components';
 
-import { Article, Content } from '../components/Main/styles';
+import { Article, Content } from '../components';
 import { ILayout } from '../types';
 
+import { Aside } from './Aside';
 import { useCreateSubMenu } from './useCreateSubmenu';
 
 import { ISubHeaderElement } from '@/types/Layout';
 import Link from 'next/link';
 import React, { FC, ReactNode } from 'react';
-
-const Aside = styled('aside', {
-  display: 'none',
-  width: '25%',
-  '@md': {
-    display: 'block',
-  },
-});
 
 export const Full: FC<ILayout> = ({ children }) => {
   const { docRef, headers } = useCreateSubMenu();
@@ -38,6 +31,7 @@ export const Full: FC<ILayout> = ({ children }) => {
         <Article ref={docRef}>{children}</Article>
 
         <Aside>
+          <Heading as="h5">On this page</Heading>
           <ul>{headers.map(renderListItem)}</ul>
         </Aside>
       </Content>
