+++ packages/apps/kadena-docs/src/components/Layout/components/index.ts
@@ -2,3 +2,4 @@ export * from './Main';
 export * from './Header';
 export * from './Footer';
 export * from './DocsLogo';
+export * from './styles';
